Move `check_if_connection_exists` to `load_file`
This moves `check_if_connection_exists` in `python-sdk/src/astro/sql/operators/load_file.py` which is the only place it is ever used
