+++ python-sdk/src/astro/files/__init__.py
@@ -1,24 +1,10 @@
 from typing import TYPE_CHECKING
 
-from airflow.hooks.base import BaseHook
-
 from astro.files.base import File, resolve_file_path_pattern  # noqa: F401 # skipcq: PY-W2000
 from astro.files.operators.files import get_file_list_func  # noqa: F401 # skipcq: PY-W2000
 
 if TYPE_CHECKING:
     from airflow.models.xcom_arg import XComArg
-
-
-def check_if_connection_exists(conn_id: str) -> bool:
-    """
-    Given an Airflow connection ID, identify if it exists.
-    Return True if it does or raise an AirflowNotFoundException exception if it does not.
-
-    :param conn_id: Airflow connection ID
-    :return bool: If the connection exists, return True
-    """
-    BaseHook.get_connection(conn_id)
-    return True
 
 
 def get_file_list(path: str, conn_id: str, **kwargs) -> "XComArg":
