+++ python-sdk/src/astro/sql/operators/load_file.py
@@ -4,6 +4,7 @@ from typing import Any
 
 import pandas as pd
 from airflow.decorators.base import get_unique_task_id
+from airflow.hooks.base import BaseHook
 from airflow.models.xcom_arg import XComArg
 
 from astro.airflow.datasets import kwargs_with_datasets
@@ -11,7 +12,7 @@ from astro.constants import DEFAULT_CHUNK_SIZE, ColumnCapitalization, LoadExistS
 from astro.databases import create_database
 from astro.databases.base import BaseDatabase
 from astro.dataframes.pandas import PandasDataframe
-from astro.files import File, check_if_connection_exists, resolve_file_path_pattern
+from astro.files import File, resolve_file_path_pattern
 from astro.settings import LOAD_FILE_ENABLE_NATIVE_FALLBACK
 from astro.sql.operators.base_operator import AstroSQLBaseOperator
 from astro.table import BaseTable
@@ -327,3 +328,15 @@ def load_file(
         enable_native_fallback=enable_native_fallback,
         **kwargs,
     ).output
+
+
+def check_if_connection_exists(conn_id: str) -> bool:
+    """
+    Given an Airflow connection ID, identify if it exists.
+    Return True if it does or raise an AirflowNotFoundException exception if it does not.
+
+    :param conn_id: Airflow connection ID
+    :return bool: If the connection exists, return True
+    """
+    BaseHook.get_connection(conn_id)
+    return True
